Add teams.ts class file
