+++ src/class/projects.ts
@@ -9,24 +9,6 @@ export type ProjectType =
   | "Mixed-use"
   | "Industrial"
   | "Heavy civil";
-export type TeamRole =
-  | "BIM Manager"
-  | "Structural"
-  | "MEP"
-  | "Architect"
-  | "Contractor";
-
-// Define the structure for a team
-export interface ITeam {
-  teamName: string;
-  teamRole: TeamRole;
-  teamDescription: string;
-  contactName: string;
-  contactPhone: string;
-  teamProjectId: string;
-  fragmentMap?: OBC.FragmentIdMap;
-  camera?: { position: THREE.Vector3; target: THREE.Vector3 };
-}
 
 // New created projects
 export interface IProject {
@@ -52,22 +34,6 @@ export function toggleModal(id: string) {
     } else modal.showModal();
   } else {
     console.warn("The provided modal wasn't found. ID: ", id);
-  }
-}
-
-// DOM element to display team information
-const teamInfo = document.getElementById("team-info") as HTMLElement;
-
-// Function to update team information in the UI
-function updateTeamInfo(team: Team) {
-  if (team) {
-    teamInfo.innerHTML = `
-      <p>Company in charge: ${team.teamName}</p>
-      <p>Company's role: ${team.teamRole}</p>
-      <p>Description: ${team.teamDescription}</p>
-      <p>Contact Name: ${team.contactName}</p>
-      <p>Phone number: ${team.contactPhone}</p>
-      <p>Project: Arreglar esto !!!</p>`;
   }
 }
 
@@ -106,27 +72,5 @@ export class Project implements IProject {
         this[key] = data[key];
       }
     }
-  }
-}
-
-// Class representing a team
-export class Team implements ITeam {
-  // Properties to satisfy ITeam
-  teamName: string;
-  teamRole: "BIM Manager" | "Structural" | "MEP" | "Architect" | "Contractor";
-  teamDescription: string;
-  contactName: string;
-  contactPhone: string;
-  teamProjectId: string;
-
-  id: string;
-  fragmentMap?: OBC.FragmentIdMap;
-  camera?: { position: THREE.Vector3; target: THREE.Vector3 };
-
-  constructor(data: ITeam, id = uuidv4()) {
-    for (const key in data) {
-      this[key] = data[key];
-    }
-    this.id = id;
   }
 }
