+++ src/react-components/DetailsPageHeader.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
-import { IProject, ITeam, Project } from "../class/projects";
+import { IProject, Project } from "../class/projects";
+import { ITeam } from "../class/teams";
 import { ProjectsManager } from "../class/projectsManager";
 
 interface Props {
