+++ src/react-components/TeamsCard.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 import * as OBC from "openbim-components";
 import * as THREE from "three";
-import { ITeam, Project, Team, TeamRole, toggleModal } from "../class/projects";
+import { Project, toggleModal } from "../class/projects";
+import { ITeam, Team, TeamRole } from "../class/teams";
 import { ProjectsManager } from "../class/projectsManager";
 // import * as Firestore from "firebase/firestore";
 // import { getCollection } from "../firebase";
